requirements.txt
fastapi[standard]==0.116.1
uvicorn==0.32.1
starlette==0.41.3
pydantic==2.10.0
srt==3.5.3
pyjwt
python-decouple==3.8
passlib==1.7.4
bcrypt==4.0.1
pydantic-csv==0.1.0
psycopg[binary,pool]==3.2.10
pandas==2.3.2
redis==6.4.0
# hiredis==3.2.1 # faster redis
python-socketio==5.13.0
simplejson==3.20.2
rq==2.6.0
kiteconnect==4.1.0
flatten_json==0.1.14
psutil==7.2.1
growwapi==1.5.0
pyotp==2.9.0
prometheus-client==0.24.1
scipy==1.17.0